Fix progress callback type and add chapters debug logging
- Fixed progress callback type mismatch (int64_t -> float)
- Added debug logging for chapters parsing to trace issues
- Unified mediaObj extraction for both chapters and audio tracks
- This helps diagnose why chapters show as 0 when tracks are found

+++ src/view/media_detail_view.cpp
@@ -534,9 +534,10 @@ void MediaDetailView::onDownload() {
 
         // Set progress callback
         DownloadsManager::getInstance().setProgressCallback(
-            [progressDialog](int64_t downloaded, int64_t total) {
+            [progressDialog](float downloaded, float total) {
                 brls::sync([progressDialog, downloaded, total]() {
-                    progressDialog->updateDownloadProgress(downloaded, total);
+                    progressDialog->updateDownloadProgress(static_cast<int64_t>(downloaded),
+                                                           static_cast<int64_t>(total));
                 });
             }
         );
